Improve performance of Option List View by no recreating the item views everytime
