+++ Sources/StreamChatUI/ChatMessageList/Attachments/Poll/PollAttachmentOptionListView/PollAttachmentOptionListItemView.swift
@@ -134,45 +134,3 @@ open class PollAttachmentOptionListItemView: _View, ThemeProvider {
         onOptionTap?(option)
     }
 }
-
-///
-open class StackedUserAvatarsView: _View, ThemeProvider {
-    struct Content {
-        var users: [ChatUser]
-    }
-
-    var content: Content? {
-        didSet {
-            updateContentIfNeeded()
-        }
-    }
-
-    open private(set) lazy var mainContainer: UIStackView = HContainer(spacing: -4)
-
-    open var userAvatarViews: [ChatUserAvatarView] {
-        guard let content = self.content else { return [] }
-        return content.users.map { user in
-            let avatarView = components.userAvatarView.init()
-                .width(20)
-                .height(20)
-            avatarView.shouldShowOnlineIndicator = false
-            avatarView.content = user
-            return avatarView
-        }
-    }
-
-    override open func setUpLayout() {
-        super.setUpLayout()
-
-        mainContainer.embed(in: self)
-    }
-
-    override open func updateContent() {
-        super.updateContent()
-
-        mainContainer.removeAllArrangedSubviews()
-        userAvatarViews.forEach {
-            mainContainer.addArrangedSubview($0)
-        }
-    }
-}
