+++ Sources/StreamChatUI/ChatMessageList/Attachments/Poll/PollAttachmentOptionListView/PollAttachmentOptionListView.swift
@@ -7,6 +7,8 @@ import UIKit
 
 /// The options list view of the poll attachment.
 open class PollAttachmentOptionListView: _View, ThemeProvider {
+    // MARK: - Content
+
     public struct Content: Equatable {
         public var poll: Poll
 
@@ -21,38 +23,56 @@ open class PollAttachmentOptionListView: _View, ThemeProvider {
         }
     }
 
-    /// The container responsible to render each option in a vertical stack.
-    /// Whenever the content changes, the stack view is rebuilt.
-    open var container: UIStackView? {
-        didSet {
-            oldValue?.removeFromSuperview()
-        }
-    }
+    // MARK: - Configuration
+
+    public var maximumNumberOfVisibleOptions = 10
 
     /// A closure that is triggered whenever the option is tapped either from the button or the item itself.
     public var onOptionTap: ((PollOption) -> Void)?
 
-    override open func updateContent() {
-        super.updateContent()
+    // MARK: - Views
 
-        container = VContainer(spacing: 24) {
-            itemViews
-        }.embed(in: self)
-    }
-
-    /// The option item views.
-    open var itemViews: [PollAttachmentOptionListItemView] {
-        guard let content = self.content else { return [] }
-        return content.poll.options.map { option in
-            let view = components.pollAttachmentOptionListItemView.init()
-            view.content = .init(
-                option: option,
-                isVotedByCurrentUser: content.poll.hasCurrentUserVoted(forOption: option)
-            )
+    /// The item views that display each option.
+    ///
+    /// The number of views is constant dependent on `maximumNumberOfVisibleOptions`.
+    /// This is to make sure views are not re-created dependent on the content.
+    /// Hiding/Showing views has better performance than re-creating the views from scratch.
+    open lazy var itemViews: [PollAttachmentOptionListItemView] = {
+        (0...maximumNumberOfVisibleOptions - 1).map { _ in
+            let view = self.components.pollAttachmentOptionListItemView.init()
             view.onOptionTap = { option in
                 self.onOptionTap?(option)
             }
             return view
+        }
+    }()
+
+    // MARK: - Lifecycle
+
+    override open func setUpLayout() {
+        super.setUpLayout()
+
+        VContainer(spacing: 24) {
+            itemViews
+        }.embed(in: self)
+    }
+
+    override open func updateContent() {
+        super.updateContent()
+
+        guard let content = self.content else {
+            return
+        }
+
+        itemViews.forEach {
+            $0.isHidden = true
+        }
+        zip(itemViews, content.poll.options).forEach { itemView, option in
+            itemView.content = .init(
+                option: option,
+                isVotedByCurrentUser: content.poll.hasCurrentUserVoted(forOption: option)
+            )
+            itemView.isHidden = false
         }
     }
 }
